fix(rrhh): anular los bonos al quedar inactivo, no solo al egresar
El commit anterior colgaba la anulacion de egresar(), pero ese boton no
es el camino que RRHH usa: dan de baja pagando la liquidacion final, que
marca activo=false llamando derecho a funcionarioService.save()
(LiquidacionFinalService:623 y :725, este ultimo por tesoreria).
generarBorrador ni siquiera exige que el funcionario este egresado. Por
ahi los bonos seguian sin anularse.

Los cuatro caminos de baja -- egresar(), el pago del finiquito, el pago
por tesoreria y el toggle del legajo -- terminan en el mismo save(), que
ya detecta la transicion para cascadear usuario y cliente. Ahi se publica
FuncionarioInactivadoEvent y un listener de rrhh hace la cancelacion, con
lo que queda cubierto tambien lo que se agregue despues.

Va con @EventListener sincrono y no con @TransactionalEventListener
AFTER_COMMIT como los de fmc/listener: esto es parte de la baja, y si el
pago del finiquito revierte, los bonos no pueden quedar anulados.

El evento viaja con el id y no con la entidad: quien escucha relee lo
suyo, no un funcionario a medio guardar. La dependencia queda en la
direccion que ya tiene el proyecto -- rrhh mirando a personas.

Cierra #295.
